# openapi-schema.yaml
openapi: 3.0.3
info:
  title: Retano API
  version: 1.0.0
  description: Retano360 — Customer Retention & Campaign Management Platform. All
    endpoints require JWT Bearer authentication unless stated otherwise.
paths:
...
  /api/v1/auth/otp/request/:
    post:
      operationId: auth_otp_request_create
      description: |-
        No authentication required. Sends a 6-digit code to the given phone number via the configured SMS provider (sms.ir by default; see OTP_PROVIDER). Rate-limited two ways: a per-phone resend cooldown (independent of DRF's own throttle_scope= 'otp_request' 5/hour limit) and a max-attempts counter enforced at verify time.

        debug_code is included in the response ONLY when the server is running with OTP_FAKE_MODE=True (local development only) — never rely on its presence.
      summary: Request an OTP code
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OTPRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OTPRequestRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTPRequestResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
...
  /api/v1/auth/register/:
    post:
      operationId: auth_register_create
      description: POST /api/v1/auth/register/
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
          description: ''
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: |-
        The standard error envelope produced by
        core.exceptions.custom_exception_handler for every RAISED DRF
        exception (ValidationError, NotAuthenticated, PermissionDenied,
        Http404, unhandled 500s, and any APIException subclass such as
        core.exceptions.OTPError / BusinessLogicError / TenantPermissionError).

        Attach this to the 400 / 401 / 403 / 404 / 409 / 429 / 500 responses
        of any endpoint whose errors are raised as exceptions rather than
        hand-built as a Response() (see individual view docstrings in
        core/schema.py for the handful of endpoints that do NOT go through
        this handler — those are documented with their own, different,
        actual shape instead of this one).
      properties:
        error:
          type: boolean
          default: true
          description: Always true on an error response.
        status_code:
          type: integer
          description: Same value as the HTTP status code of this response.
        message:
          type: string
          description: Human-readable summary of what went wrong.
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: 'Field-level validation errors, e.g. {"phone_number": ["This
            field is required."]}. Empty object ({}) when the error has no field-level
            breakdown (e.g. authentication/permission/404 errors).'
      required:
      - message
      - status_code
...
    OTPRequestRequest:
      type: object
      description: POST /api/v1/auth/otp/request/ — accepts a phone number to OTP.
      properties:
        phone_number:
          type: string
          minLength: 1
      required:
      - phone_number
    OTPRequestResponse:
      type: object
      description: |-
        200 response for POST /api/v1/auth/otp/request/.

        debug_code is ONLY ever present when the server is running with
        OTP_FAKE_MODE=True (see config/settings/development.py) — it is never
        sent by a production deployment. Documented as not required so the
        schema reflects that a client must not assume it will be there.
      properties:
        phone_number:
          type: string
          description: Normalized E.164 form, e.g. +989121234567.
        ttl_seconds:
          type: integer
          description: How long the OTP code remains valid.
        resend_in_seconds:
          type: integer
          description: Client must wait this long before requesting another OTP for
            this number.
        debug_code:
          type: string
          description: DEV/TEST ONLY. Present only when the server has OTP_FAKE_MODE=True.
            Never present in production.
      required:
      - phone_number
      - resend_in_seconds
      - ttl_seconds
...
    Register:
      type: object
      description: |-
        POST /api/v1/auth/register/

        Creates a CustomUser from a phone number. Optional profile fields may
        be supplied up front; everything else is filled in later via
        /api/v1/profile/. Saving triggers the existing post_save signal on
        CustomUser, which creates the user's Tenant.
      properties:
        phone_number:
          type: string
        first_name:
          type: string
          nullable: true
          maxLength: 20
        last_name:
          type: string
          nullable: true
          maxLength: 20
        shop_name:
          type: string
          maxLength: 255
      required:
      - phone_number
    RegisterRequest:
      type: object
      description: |-
        POST /api/v1/auth/register/

        Creates a CustomUser from a phone number. Optional profile fields may
        be supplied up front; everything else is filled in later via
        /api/v1/profile/. Saving triggers the existing post_save signal on
        CustomUser, which creates the user's Tenant.
      properties:
        phone_number:
          type: string
          minLength: 1
        first_name:
          type: string
          nullable: true
          maxLength: 20
        last_name:
          type: string
          nullable: true
          maxLength: 20
        shop_name:
          type: string
          minLength: 1
          maxLength: 255
      required:
      - phone_number
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    syncApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: Tenant Sync API Key
      description: 'Per-tenant sync API key for the ETL-facing endpoints under /api/v1/sync/
        (config fetch, data ingest, run reporting). This is NOT a JWT and is unrelated
        to the jwtAuth scheme used by every other endpoint in this API. Generate a
        key from the human-facing POST /api/v1/sync-conf/generate-key/ endpoint (JWT-authenticated),
        then supply it here as: Authorization: Bearer <api_key>. The key is enabled
        only after every field mapping row has been completed on the API-Conf page;
        a disabled or unknown key returns 401.'